UN-3607 [FIX] greptile — unify the PG poll loop (move poll_for_row to unstract.core)
greptile P2: poll_for_row was backend-only — the workers PgResultBackend.wait_for_result
still had its own identical backoff loop, so the constants lived in two places. Since
poll_for_row is pure (no Django/psycopg/SDK deps), moved it to unstract.core.polling
where BOTH PG result pollers import it: the backend DjangoQueueTransport and the workers
PgResultBackend. The backoff now lives in exactly one place to tune.

- new unstract.core.polling.poll_for_row (the shared backoff skeleton)
- workflow_execution.executor_rpc: dropped poll_for_row (+ its time/TypeVar imports)
- backend adapter + PgResultBackend.wait_for_result both delegate to it
- behaviour-identical: backend 8 + workers 38 + result_backend 7 (real-PG) green
